tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/animation-play-state_paused.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - animation-play-state: paused</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#the-animation-play-state-property-">
   <meta name="flags" content="dom">
   <meta name="assert" content="A running animation can be paused by setting
   'animation-play-state' to 'paused'">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes grow {
         from {
            width: 50px;
         }
         to {
            width: 150px;
         }
      }
      div {
         width: 50px;
         height: 100px;
         background: blue;
         animation-duration: 1s;
         animation-timing-function: steps(2);
      }
      .test {
         animation-name: grow;
      }
      .paused {
         animation-play-state: paused;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var style = getComputedStyle(div, null);
         var passed = false;
         setTimeout(function() {
            div.className = 'test paused';
            passed = (style.width == '100px');
            setTimeout(function() {
               passed = (passed && style.width == '100px');
               test(function(){assert_true(passed);}, testName); done();
               log.data = (passed ? 'PASS if the animation occurred.' : 'FAIL');
            }, 300);
         }, 750);
         div.className = 'test';
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The blue block below should become a square, and stay that way.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
